+++ src/macros.rs
@@ -737,3 +737,23 @@ macro_rules! ei_object_base {
         }
     };
 }
+
+macro_rules! logical_to_client_wire_scale {
+    ($client:expr, $($field:expr),+ $(,)?) => {
+        if let Some(scale) = $client.wire_scale.get() {
+            $(
+                $field = $field * scale;
+            )+
+        }
+    };
+}
+
+macro_rules! client_wire_scale_to_logical {
+    ($client:expr, $($field:expr),+ $(,)?) => {
+        if let Some(scale) = $client.wire_scale.get() {
+            $(
+                $field = $field / scale;
+            )+
+        }
+    };
+}